cambio en los campos de las formas

+++ resources/views/Traficos/TraficosModificar.blade.php
@@ -245,7 +245,7 @@
                                   message: 'El costo debe tener como máximo 255 caracteres'
                               },
 
-                              integer:{
+                              numeric:{
                                   message: 'Ingresa un número'
                               }
                           }
